+++ Battlefield_test.cbp
@@ -2,22 +2,26 @@
 <CodeBlocks_project_file>
 	<FileVersion major="1" minor="6" />
 	<Project>
-		<Option title="Lazyfoo14" />
+		<Option title="Battlefield_test" />
 		<Option pch_mode="2" />
 		<Option compiler="gcc" />
 		<Build>
 			<Target title="Debug">
-				<Option output="bin/Debug/Lazyfoo14" prefix_auto="1" extension_auto="1" />
+				<Option output="bin/Debug/Battlefield_test" prefix_auto="1" extension_auto="1" />
 				<Option object_output="obj/Debug/" />
 				<Option type="1" />
 				<Option compiler="gcc" />
 				<Compiler>
 					<Add option="-g" />
+					<Add directory="C:/SDL2-2.0.9/i686-w64-mingw32/include/SDL2" />
 					<Add directory="include" />
 				</Compiler>
+				<Linker>
+					<Add directory="C:/SDL2-2.0.9/i686-w64-mingw32/lib" />
+				</Linker>
 			</Target>
 			<Target title="Release">
-				<Option output="bin/Release/Lazyfoo14" prefix_auto="1" extension_auto="1" />
+				<Option output="bin/Release/Battlefield_test" prefix_auto="1" extension_auto="1" />
 				<Option object_output="obj/Release/" />
 				<Option type="1" />
 				<Option compiler="gcc" />
@@ -44,15 +48,14 @@
 		<Unit filename="include/LTexture.h" />
 		<Unit filename="include/base.h" />
 		<Unit filename="include/loop.h" />
-		<Unit filename="include/soldier_short.h" />
+		<Unit filename="include/men_with_arms.h" />
 		<Unit filename="main.cpp" />
 		<Unit filename="src/BATTLE_SCENE.cpp" />
 		<Unit filename="src/LButton.cpp" />
 		<Unit filename="src/LTexture.cpp" />
 		<Unit filename="src/base.cpp" />
 		<Unit filename="src/loop.cpp" />
-		<Unit filename="src/soldier_short.cpp" />
-		<Unit filename="src/tess.cpp" />
+		<Unit filename="src/men_with_arms.cpp" />
 		<Extensions>
 			<code_completion />
 			<envvars />
