+++ Battlefield_test.depend
@@ -1,0 +1,86 @@
+# depslib dependency file v1.0
+1545552261 source:c:\users\user\desktop\battlefield_test\battlefield_test\src\base.cpp
+	"base.h"
+
+1545552596 c:\users\user\desktop\battlefield_test\battlefield_test\include\base.h
+
+1545552206 source:c:\users\user\desktop\battlefield_test\battlefield_test\src\lbutton.cpp
+	"LButton.h"
+
+1545552596 c:\users\user\desktop\battlefield_test\battlefield_test\include\lbutton.h
+	<SDL2/SDL.h>
+	<SDL_image.h>
+	<stdio.h>
+	<string>
+	<cmath>
+
+1543746997 c:\program files (x86)\sdl2\i686-w64-mingw32\include\sdl2\sdl_image.h
+	"SDL2/SDL.h"
+	"SDL2/SDL_version.h"
+	"SDL2/begin_code.h"
+	"SDL2/close_code.h"
+
+1545552332 source:c:\users\user\desktop\battlefield_test\battlefield_test\src\loop.cpp
+	"loop.h"
+
+1545552596 c:\users\user\desktop\battlefield_test\battlefield_test\include\loop.h
+	<SDL2/SDL.h>
+	<SDL_image.h>
+	<LButton.h>
+	<LTexture.h>
+	<base.h>
+
+1545552596 c:\users\user\desktop\battlefield_test\battlefield_test\include\ltexture.h
+	<SDL2/SDL.h>
+	<SDL_image.h>
+	<stdio.h>
+	<string>
+	<cmath>
+
+1545552596 source:c:\users\user\desktop\battlefield_test\battlefield_test\src\ltexture.cpp
+	"LTexture.h"
+
+1545552372 source:c:\users\user\desktop\battlefield_test\battlefield_test\src\men_with_arms.cpp
+	"men_with_arms.h"
+
+1545552961 c:\users\user\desktop\battlefield_test\battlefield_test\include\men_with_arms.h
+	<stdio.h>
+	<SDL2/SDL.h>
+	<SDL_image.h>
+	<LButton.h>
+	<LTexture.h>
+	<base.h>
+
+1545552604 source:c:\users\user\desktop\battlefield_test\battlefield_test\main.cpp
+	<SDL2/SDL.h>
+	<SDL_image.h>
+	<stdio.h>
+	<ostream>
+	<string>
+	<cmath>
+	<loop.h>
+	<LTexture.h>
+	<LButton.h>
+	<men_with_arms.h>
+	<base.h>
+	<BATTLE_SCENE.h>
+
+1545552987 c:\users\user\desktop\battlefield_test\battlefield_test\include\battle_scene.h
+	<iostream>
+	<SDL2/SDL.h>
+	<SDL_image.h>
+	<stdio.h>
+	<string.h>
+	<cmath>
+	<loop.h>
+	<LTexture.h>
+	<LButton.h>
+	<men_with_arms.h>
+	<base.h>
+	<fstream>
+	<sstream>
+	<stdlib.h>
+
+1545552596 source:c:\users\user\desktop\battlefield_test\battlefield_test\src\battle_scene.cpp
+	"BATTLE_SCENE.h"
+
