--- Lazyfoo14.depend
@@ -1,59 +1,0 @@
-# depslib dependency file v1.0
-1544437417 source:c:\users\user\desktop\sdltest\lazyfoo14\main.cpp
-	<SDL2/SDL.h>
-	<SDL_image.h>
-	<stdio.h>
-	<string>
-
-1543746997 c:\program files (x86)\sdl2\i686-w64-mingw32\include\sdl2\sdl_image.h
-	"SDL2/SDL.h"
-	"SDL2/SDL_version.h"
-	"SDL2/begin_code.h"
-	"SDL2/close_code.h"
-
-1544584863 source:c:\users\user\desktop\battlefield\lazyfoo14\main.cpp
-	<SDL2/SDL.h>
-	<SDL_image.h>
-	<stdio.h>
-	<string>
-
-1544629941 source:c:\users\user\desktop\battlefield - bigger\code\main.cpp
-	<SDL2/SDL.h>
-	<SDL_image.h>
-	<stdio.h>
-	<string>
-	<cmath>
-	<loop.h>
-	<LTexture.h>
-	<soldier_short.h>
-
-1544628088 source:c:\users\user\desktop\battlefield - bigger\code\src\loop.cpp
-	"loop.h"
-
-1544632137 c:\users\user\desktop\battlefield - bigger\code\include\loop.h
-
-1544628269 source:c:\users\user\desktop\battlefield - bigger\code\src\soldier_short.cpp
-	"soldier_short.h"
-
-1544628308 c:\users\user\desktop\battlefield - bigger\code\include\soldier_short.h
-
-1544630277 source:c:\users\user\desktop\battlefield - bigger\code\src\ltexture.cpp
-	"LTexture.h"
-
-1544630277 c:\users\user\desktop\battlefield - bigger\code\include\ltexture.h
-	<SDL2/SDL.h>
-	<SDL_image.h>
-	<stdio.h>
-	<string>
-	<cmath>
-
-1544631159 source:c:\users\user\desktop\battlefield - bigger\code\src\lbutton.cpp
-	"LButton.h"
-
-1544632362 c:\users\user\desktop\battlefield - bigger\code\include\lbutton.h
-	<SDL2/SDL.h>
-	<SDL_image.h>
-	<stdio.h>
-	<string>
-	<cmath>
-
