+++ image/character/duck_green.png
