+++ property.csv
@@ -1,2 +1,2 @@
-﻿0,DUCK_ORI,0,100,20,1,1,stand.png
-1,DUCK_GREEN,1,120,23,1,2,stand2.png
+﻿0,DUCK_ORI,0,100,20,1,1,image/character/duck_ori.png
+1,DUCK_GREEN,1,120,23,1,2,image/character/duck_green.png
