+++ src/LButton.cpp
@@ -2,10 +2,7 @@
 
 LButton::LButton()
 {
-    mPosition.x = 0;
-	mPosition.y = 0;
-
-	mCurrentSprite = BUTTON_SPRITE_MOUSE_OUT;
+    //ctor
 }
 
 LButton::~LButton()
