+++ src/men_with_arms.cpp
@@ -1,0 +1,11 @@
+#include "men_with_arms.h"
+
+men_with_arms::men_with_arms()
+{
+    //ctor
+}
+
+men_with_arms::~men_with_arms()
+{
+    //dtor
+}
