--- src/soldier_short.cpp
@@ -1,13 +1,0 @@
-#include "soldier_short.h"
-
-soldier_short::soldier_short()
-{
-    //ctor
-}
-
-soldier_short::~soldier_short()
-{
-
-        //printf("DESTRUCT%d",id);
-
-}
